fish

+++ Personal/settings/ubuntu-chadwm/install-chadwm-on-ubuntu.sh
@@ -87,10 +87,17 @@ sudo make install
 sudo apt remove -y xdg-desktop-portal-gnome
 
 
+if -f ~/.bashrc; then
+	echo '
+	alias update="sudo apt update && sudo apt upgrade"' | tee -a ~/.bashrc
+	exec bash
+fi
 
-echo '
-alias update="sudo apt update && sudo apt upgrade"' | tee -a ~/.bashrc
-exec bash
+if -f ~/.config/fish/config.fish; then
+	echo '
+	alias update="sudo apt update && sudo apt upgrade"' | tee -a ~/.config/fish/config.fish
+	exec fish
+fi
 
 echo
 tput setaf 6
